requirements.txt
caltechdata_api>=1.2.1
python-hglib
